flask_app/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==9.5.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
openpyxl==3.1.2
